Update code to use size_t and ssize_t
The`size_t` type is used to represent the size of an object.
It is defined as the return type of the sizeof operator. As an
implication size_t is used for holding an array index.

Many of the functions that would naturally return `size_t`
also return -1 when an error has occured. To hold the error
these functions have been setup to return `ssize_t`. A posix
value that is often described as `signed size_t`. If properly
defined it should only return -1 for errors all other values
should be positive. For that reason there are some functions
that would have made since to return ssize_t but due to
returning more error values than just -1 those functions are
returning long. An example of this can be found in the windows
storage.c code.

Being consistant with the usage of `size_t` and `ssize_t`
helps reduce the amount of casting when compiling for systems
that are not 32-bit.

This is a first pass through the code to identify locations
that `int` or `long` was used when `size_t` or `ssize_t`
probably should have been used.

This is not an exastive update.

+++ api/oc_client_api.c
@@ -35,7 +35,7 @@ oc_event_callback_retval_t oc_ri_remove_client_cb(void *data);
 static bool
 dispatch_coap_request(void)
 {
-  int payload_size = oc_rep_finalize();
+  ssize_t payload_size = oc_rep_finalize();
 
   if ((client_cb->method == OC_PUT || client_cb->method == OC_POST) &&
       payload_size > 0) {
