+++ port/windows/ipadapter.c
@@ -78,7 +78,7 @@ typedef struct ip_context_t
   HANDLE event_server_handle;
   DWORD event_thread;
   BOOL terminate;
-  int device;
+  size_t device;
 } ip_context_t;
 
 #ifdef OC_DYNAMIC_ALLOCATION
@@ -120,7 +120,7 @@ oc_network_event_handler_mutex_destroy(void)
 }
 
 static ip_context_t *
-get_ip_context_for_device(int device)
+get_ip_context_for_device(size_t device)
 {
 #ifdef OC_DYNAMIC_ALLOCATION
   ip_context_t *dev = oc_list_head(ip_contexts);
@@ -880,7 +880,7 @@ network_event_thread_error:
 }
 
 oc_endpoint_t *
-oc_connectivity_get_endpoints(int device)
+oc_connectivity_get_endpoints(size_t device)
 {
   ip_context_t *dev = get_ip_context_for_device(device);
   if (!dev) {
@@ -1313,7 +1313,7 @@ connectivity_ipv4_init(ip_context_t *dev)
 #endif
 
 int
-oc_connectivity_init(int device)
+oc_connectivity_init(size_t device)
 {
   if (!ifchange_initialized) {
     WSADATA wsadata;
@@ -1504,7 +1504,7 @@ oc_connectivity_init(int device)
 }
 
 void
-oc_connectivity_shutdown(int device)
+oc_connectivity_shutdown(size_t device)
 {
   ip_context_t *dev = get_ip_context_for_device(device);
   dev->terminate = TRUE;
